Extract shared state and business logic into state.py
Pulls the runtime state and business logic out of main.py into a new state
module with no FastAPI dependency, so it can be shared by the web layer and
the monitor loop without a circular import:

- Apprise object, HTTP session (+ get/cleanup), metrics, the live ID/URL
  lists and their locks/accessors.
- .env persistence (update_env_file), TestFlight ID + Apprise URL validation,
  and the add/remove helpers.
- The GitHub update checker, shutdown event/handler, and the os.execv restart.

main.py imports these (singletons shared by identity) and keeps the FastAPI
app, auth, lifespan, routes, the monitor loop, and the entrypoint. Drops the
now-unused imports from main.py. The API tests are updated to patch the moved
internals on `state`.

main.py: ~1,700 -> ~1,010 lines. Behaviour unchanged.
